requirements.txt
blinker==1.9.0
click==8.1.8
duckduckgo_search==7.2.1
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
primp==0.10.1
Werkzeug==3.1.3
